requirements.txt
openai-whisper
transformers
torch
langchain
pyaudio
speechrecognition
fastapi
uvicorn
chromadb
nltk